decouple controllers from usecases
